+++ src/stdio.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
+import { createMarketingServer } from "./index.js";
+
+await createMarketingServer().connect(new StdioServerTransport());